requirements.txt
streamlit
mistralai
python-dotenv
pandas
plotly
matplotlib
pillow
pyperclip
